+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # `nonempty-collections`
 
-## Unreleased
+## 1.1.0 (2025-10-16)
 
 #### Added